chore(deps): Update Spark dependency to 2.9.4

+++ aws-serverless-java-container-spark/src/main/java/com/amazonaws/serverless/proxy/spark/embeddedserver/LambdaEmbeddedServer.java
@@ -80,7 +80,7 @@ public class LambdaEmbeddedServer
 
 
     public void configureWebSockets(Map<String, WebSocketHandlerWrapper> webSocketHandlers,
-                                    Optional<Integer> webSocketIdleTimeoutMillis) {
+                                    Optional<Long> webSocketIdleTimeoutMillis) {
         // Swallowing this exception to prevent Spark from getting stuck
         // throw new UnsupportedOperationException();
         log.info("Spark called configureWebSockets. However, web sockets are not supported");
